Add task error handling

+++ ml_enabler/services/task_service.py
@@ -1,5 +1,7 @@
 from ml_enabler.models.task import Task
 from ml_enabler import db
+from ml_enabler.models.utils import NotFound
+import boto3
 
 class TaskService():
     @staticmethod
@@ -17,10 +19,17 @@ class TaskService():
     @staticmethod
     def delete(task_id: int):
         task = Task.get(task_id)
-        task.delete()
+        if task:
+            task.delete()
+        else:
+            raise NotFound
 
+    @staticmethod
     def list(pred_id: int, task_type: str):
         rawtasks = Task.list(pred_id, task_type)
+
+        if not rawtasks:
+            raise NotFound
 
         tasks = []
         if (rawtasks):
@@ -37,3 +46,31 @@ class TaskService():
                 'tasks': []
             }
 
+    @staticmethod
+    def get(task_id: int):
+        task = Task.get(task_id)
+
+        if not task:
+            raise NotFound
+
+        task = task.as_dto().to_primitive()
+
+        if task.get('batch_id') is not None:
+            batch = boto3.client(
+                service_name='batch',
+                region_name='us-east-1',
+                endpoint_url='https://batch.us-east-1.amazonaws.com'
+            )
+
+            desc = batch.describe_jobs(
+                jobs = [ task.get('batch_id') ]
+            )
+
+            if len(desc.get('jobs')) == 1:
+                task['status'] = desc['jobs'][0].get('status')
+                task['statusReason'] = desc['jobs'][0].get('statusReason')
+            else:
+                task['status'] = 'UNKNOWN'
+                task['statusReason'] = 'AWS does not report this task'
+
+        return task
